Add a location page that uses Google Maps API (#4)
This PR creates a new location page in the portfolio that makes use of the
Google Maps API to display the different places I have lived. It completes the
first half of the Google STEP week 4 walkthroughs. It also corrects a typo
from PR #3. Other features include HTML elements that fade in upon scroll
between maps.